+++ packages/react/objects/facets/Direction.ts
@@ -1,0 +1,6 @@
+export enum DirectionEnum {
+  ROW = 'row',
+  COLUMN = 'column',
+}
+
+export type DirectionEnumValues = `${DirectionEnum}`