+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Update SegmentControl padding
 - Update radio's padding
 - Suffix colors variables for themes
+- Add Stack component
 
 ## [4.7.4] - 2025-08-22
 - Fix iconNameRight Icon (native)